% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePipThreshold.R
\name{CalculatePipThreshold}
\alias{CalculatePipThreshold}
\title{Calculate PIP Threshold}
\usage{
CalculatePipThreshold(
  y,
  absCV,
  sampSize,
  coeffs_ls = list(A = 0, K = 1.43462, C = 1.59948, betaAbsCV = 0.56615, betaSampSize =
    0.51345),
  na.rm = TRUE
)
}
\arguments{
\item{y}{a response vector for BKMR}

\item{absCV}{If \code{y} is not supplied, the absolute value of the coefficient
of variation of the response}

\item{sampSize}{If \code{y} is not supplied, the number of observations included
in the response}

\item{coeffs_ls}{A list of Richard's Curve parameters. See Details.}

\item{na.rm}{Remove missing values from \code{y}? Defaults to \code{TRUE}}
}
\value{
A single numeric value; the output of the Richard's Four-Parameter
Logistic Regression curve with the coefficient values supplied in
\code{coeffs_ls}.
}
\description{
Given a response vector (or statistics from this vector),
calculate a PIP threshold that should preserve close to a nominal 5\% test
size for Bayesian Kernel Machine Regression (BKMR) feature selection.
}
\details{
Tanvir: include a link to the Richard's curve, LaTeX code to show
what this curve looks like, and a brief (1-2 sentence) explanation of how we
calculated the values in \code{coeffs_ls}.
}
\examples{
CalculatePipThreshold(absCV = 7.5, sampSize = 300)
# should equal 0.6829892
}
