% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateMultivariateMoments.R
\name{estimate_mv_moments}
\alias{estimate_mv_moments}
\title{Helper function to estimate moment vectors/matrices for observations within
a group}
\usage{
estimate_mv_moments(x_df)
}
\arguments{
\item{x_df}{A numeric data frame with observations from ONE group}
}
\value{
A list of statistics/moments (sample size, mean, standard deviation,
correlation matrix, skewness) as vectors/matrices
}
\description{
Helper function to estimate moment vectors/matrices for observations within
a group
}
\examples{
myData <- data.frame(
  VALUE1 = c(2.3, 2.7, 2.5),
  VALUE2 = c(4.5, 4.2, 4.35)
)
estimate_mv_moments(myData)
}
