% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metalExposChildren.R
\docType{data}
\name{metalExposChildren_df}
\alias{metalExposChildren_df}
\title{Children's Environmental Heavy Metal Exposure Data}
\format{
\subsection{\code{who}}{

A data frame with 437 rows and 13 columns:
\describe{
\item{ID}{Subject's ID}
\item{age}{Subject's age in years}
\item{QI}{Intelligence quotient}
\item{Cadmium}{Cd urine concentration	ng/ml}
\item{Mercury}{Hg urine concentration	ng/ml}
\item{Arsenic}{As urine concentration	ng/ml}
\item{Lead}{Pb blood concentration	ng/ml}
\item{Manganese}{Mn hair concentration	ng/g}
\item{Sex}{Student's gender; 1 = Male, 2 = Female}
\item{BMI}{Body Mass Index (kg/m2)}
\item{SES}{Social Economic Status; 1 = LOW, 2 = MEDIUM, 3 = HIGH}
\item{Distance_metres}{Distance to nearest industrial site (m)}
\item{SPM}{Raven's Standard Progressive Matrices}
}
}
}
\source{
Prof. Roberto Lucchini; see inst/scripts/data_metalExposChildren*.R
for more details. Licensed under CC BY 4.0.
}
\usage{
metalExposChildren_df
}
\description{
A short description...
}
\keyword{datasets}
