% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tarantoChildren.R
\docType{data}
\name{tarantoChildren_df}
\alias{tarantoChildren_df}
\title{Taranto Children's Environmental Exposure Data}
\format{
\subsection{\code{who}}{

A data frame with 437 rows and 13 columns:
\describe{
\item{ID}{}
\item{age}{}
\item{QI}{}
\item{Cadmium}{}
\item{Mercury}{}
\item{Arsenic}{}
\item{Lead}{}
\item{Manganese}{}
\item{Sex}{}
\item{BMI}{}
\item{SES}{}
\item{Distance_metres}{}
\item{SPM}{}
}
}
}
\source{
Prof. Roberto Lucchini; see inst/scripts/data_tarantoChildren*.R for
more details. Licensed under CC BY 4.0.
}
\usage{
tarantoChildren_df
}
\description{
A short description...
}
\keyword{datasets}
