% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateMultivariateMoments.R
\name{.skewness}
\alias{.skewness}
\title{Helper function to calculate skewness for a vector}
\usage{
.skewness(x, mean_vec, sampSD, N)
}
\arguments{
\item{x}{A numeric vector of data}

\item{mean_vec}{The mean of the data}

\item{sampSD}{The standard deviation of the data}

\item{N}{The sample size}
}
\value{
The skewness value
}
\description{
Helper function to calculate skewness for a vector
}
