% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateMvGammaData.R
\name{generate_mvGamma_data}
\alias{generate_mvGamma_data}
\title{Generate Multivariate Skewed Gamma Transformed Data}
\usage{
generate_mvGamma_data(N, mean_vec, cov_mat, shape_num, rate_num)
}
\arguments{
\item{N}{Number of samples to generate.}

\item{mean_vec}{A numeric vector of means for the normal distribution.}

\item{cov_mat}{A covariance matrix for the normal distribution.}

\item{shape_num}{A numeric vector of shape parameters for the Gamma transformation.}

\item{rate_num}{A numeric vector of rate parameters for the Gamma transformation.}
}
\value{
A list containing:
\itemize{
\item \code{norm_samples}: The generated multivariate normal samples.
\item \code{norm_cdf}: The CDF of the normal samples.
\item \code{gamma_samples}: The transformed Gamma samples.
}
}
\description{
This function generates multivariate normal samples, transforms them into Z-scores,
and then applies a \code{qgamma} transformation to each variable.
}
\examples{
p <- 4
N <- 1000
shapeGamma_num <- c(0.5, 0.75, 1, 1.25)
rateGamma_num <- 1:4
mean_vec <- rep(0, p)
cov_mat <- diag(p)
generate_mvGamma_data(N, mean_vec, cov_mat, shapeGamma_num, rateGamma_num)
}
